requirements.txt
langchain
groq
python-dotenv
langchain_groq
langchain-groq
grandalf